Split reconstruction.cc in smaller files
